MicroService/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.6
numpy==1.26.3
soundfile==0.12.1
torch>=2.0.0
noisereduce==3.0.0
webrtcvad==2.0.10
scipy==1.11.4
python-dotenv==1.0.0
openai-whisper
